# internal/app/feedback/handler/feedback_handler.go
package handler

import (
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
	"github.com/nathakusuma/conference-backend/domain/contract"
	"github.com/nathakusuma/conference-backend/domain/dto"
	"github.com/nathakusuma/conference-backend/domain/enum"
	"github.com/nathakusuma/conference-backend/domain/errorpkg"
	"github.com/nathakusuma/conference-backend/internal/middleware"
	"github.com/nathakusuma/conference-backend/pkg/validator"
)

type feedbackHandler struct {
	svc contract.IFeedbackService
	val validator.IValidator
}

func InitFeedbackHandler(
	router fiber.Router,
	midw *middleware.Middleware,
	val validator.IValidator,
	feedbackSvc contract.IFeedbackService,

) {
	handler := feedbackHandler{
		svc: feedbackSvc,
		val: val,
	}

	feedbackGroup := router.Group("/feedbacks")
	feedbackGroup.Use(midw.RequireAuthenticated())

	feedbackGroup.Post("",
		midw.RequireOneOfRoles(enum.RoleUser),
		handler.createFeedback(),
	)

	feedbackGroup.Get("/conferences/:id",
		handler.getFeedbacksByConferenceID(),
	)

	feedbackGroup.Delete("/:id",
		midw.RequireOneOfRoles(enum.RoleEventCoordinator),
		handler.deleteFeedback(),
	)
}

func (h *feedbackHandler) createFeedback() fiber.Handler {
	return func(ctx *fiber.Ctx) error {
		type request struct {
			ConferenceID uuid.UUID `json:"conference_id" validate:"required,uuid"`
			Comment      string    `json:"comment" validate:"required,min=3,max=1000"`
		}

		var req request
		if err := ctx.BodyParser(&req); err != nil {
			return errorpkg.ErrFailParseRequest
		}

		if err := h.val.ValidateStruct(req); err != nil {
			return err
		}

		userID, _ := ctx.Locals("user.id").(uuid.UUID)

		feedbackID, err := h.svc.CreateFeedback(ctx.Context(), userID, req.ConferenceID, req.Comment)
		if err != nil {
			return err
		}

		return ctx.Status(fiber.StatusCreated).JSON(fiber.Map{
			"feedback": dto.FeedbackResponse{ID: feedbackID},
		})
	}
}

func (h *feedbackHandler) getFeedbacksByConferenceID() fiber.Handler {
	return func(c *fiber.Ctx) error {
		conferenceID, err := uuid.Parse(c.Params("id"))
		if err != nil {
			return errorpkg.ErrFailParseRequest
		}

		var lazyReq dto.LazyLoadQuery
		if err := c.QueryParser(&lazyReq); err != nil {
			return errorpkg.ErrFailParseRequest
		}

		if err := h.val.ValidateStruct(lazyReq); err != nil {
			return err
		}

		feedbacks, lazyResp, err := h.svc.GetFeedbacksByConferenceID(c.Context(), conferenceID, lazyReq)
		if err != nil {
			return err
		}

		return c.JSON(map[string]interface{}{
			"feedbacks":  feedbacks,
			"pagination": lazyResp,
		})
	}
}

func (h *feedbackHandler) deleteFeedback() fiber.Handler {
	return func(c *fiber.Ctx) error {
		feedbackID, err := uuid.Parse(c.Params("id"))
		if err != nil {
			return errorpkg.ErrFailParseRequest
		}

		if err := h.svc.DeleteFeedback(c.Context(), feedbackID); err != nil {
			return err
		}

		return c.SendStatus(fiber.StatusNoContent)
	}
}